Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.5",
    "google-adk>=1.24.1",
    "pydantic",
    "python-dotenv",
    "manim",
    "nest-asyncio>=1.6.0",
    "numpy>=2.4.2",
]
